scan_sources: move command from __main__ to separate module
